fix: Be able to connect to SQL in containers from CMS
By default if you use `localhost` as the `HOST` value for mysql, it
tries to connect to a file socket on disk rather than trying to connect
to the loopback hostname.  This prevents us from running MySQL in a
container while running the LMS on your local machine.

Setting the host to `127.0.0.1` forces the SQL connection to go over TCP
instead. This allows you to map your container port to your localhost
without any issues.

We did this in lms/envs/common.py in an earlier change but did not
update cms/envs/common.py at that time.